refactor runMotor to runOutput
